+++ src/main/java/gaji/service/aws/s3/AmazonS3Manager.java
@@ -7,8 +7,8 @@ import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import gaji.service.config.AmazonConfig;
-import gaji.service.global.common.exception.RestApiException;
-import gaji.service.global.common.exception.code.status.GlobalErrorStatus;
+import gaji.service.global.exception.RestApiException;
+import gaji.service.global.exception.code.status.GlobalErrorStatus;
 import gaji.service.global.util.uuid.Uuid;
 import gaji.service.global.util.uuid.UuidRepository;
 import lombok.RequiredArgsConstructor;
