+++ src/main/java/gaji/service/domain/file/controller/FileController.java
@@ -3,8 +3,8 @@ package gaji.service.domain.file.controller;
 import gaji.service.domain.file.dto.response.FileCreateResponse;
 import gaji.service.domain.file.dto.response.FileDeleteResponse;
 import gaji.service.domain.file.service.FileService;
-import gaji.service.global.common.base.BaseResponse;
-import gaji.service.global.common.enums.FileCategory;
+import gaji.service.global.base.BaseResponse;
+import gaji.service.global.enums.FileCategory;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
