+++ src/main/java/gaji/service/domain/file/service/FileService.java
@@ -2,7 +2,7 @@ package gaji.service.domain.file.service;
 
 import gaji.service.domain.file.dto.response.FileCreateResponse;
 import gaji.service.domain.file.dto.response.FileDeleteResponse;
-import gaji.service.global.common.enums.FileCategory;
+import gaji.service.global.enums.FileCategory;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface FileService {
