+++ src/main/java/gaji/service/domain/file/service/FileServiceImpl.java
@@ -6,9 +6,9 @@ import gaji.service.domain.file.dto.response.FileCreateResponse;
 import gaji.service.domain.file.dto.response.FileDeleteResponse;
 import gaji.service.domain.file.mapper.FileMapper;
 import gaji.service.domain.file.repository.FileRepository;
-import gaji.service.global.common.enums.FileCategory;
-import gaji.service.global.common.exception.RestApiException;
-import gaji.service.global.common.exception.code.status.FilesErrorStatus;
+import gaji.service.global.enums.FileCategory;
+import gaji.service.global.exception.RestApiException;
+import gaji.service.global.exception.code.status.FilesErrorStatus;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
