+++ src/main/java/gaji/service/global/exception/code/status/GlobalErrorStatus.java
@@ -1,8 +1,8 @@
-package gaji.service.global.common.exception.code.status;
+package gaji.service.global.exception.code.status;
 
 
-import gaji.service.global.common.exception.code.BaseCodeDto;
-import gaji.service.global.common.exception.code.BaseErrorCodeInterface;
+import gaji.service.global.exception.code.BaseCodeDto;
+import gaji.service.global.exception.code.BaseErrorCodeInterface;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
